Completed Bridge System, first Revision. Moved IrcEndPoint and DiscordEndPoint to the net.dv8tion.discord.endpoint.types package. Also renamed IRCConnectInfo -> IrcConnectInfo.
