+++ src/net/dv8tion/discord/bridge/endpoint/EndPointManager.java
@@ -3,6 +3,9 @@ package net.dv8tion.discord.bridge.endpoint;
 import java.util.ArrayList;
 import java.util.List;
 
+import net.dv8tion.discord.bridge.endpoint.types.DiscordEndPoint;
+import net.dv8tion.discord.bridge.endpoint.types.IrcEndPoint;
+
 public class EndPointManager
 {
     private static EndPointManager manager;
