+++ src/net/dv8tion/discord/bridge/endpoint/types/IrcEndPoint.java
@@ -1,6 +1,10 @@
-package net.dv8tion.discord.bridge.endpoint;
+package net.dv8tion.discord.bridge.endpoint.types;
 
 import net.dv8tion.discord.Bot;
+import net.dv8tion.discord.bridge.endpoint.EndPoint;
+import net.dv8tion.discord.bridge.endpoint.EndPointInfo;
+import net.dv8tion.discord.bridge.endpoint.EndPointMessage;
+import net.dv8tion.discord.bridge.endpoint.EndPointType;
 
 import org.pircbotx.Channel;
